put 'useMock' in separate file and expand
